Embed the license to file to make drop-in easier (#14)
This library is distributed with MIT license. It requires to show the
copyright notice in software. So when embedding this library in a plugin,
a user also needs to show the copyright in the plugin.

By embedding a license description and copyright notice to source file,
a user doesn't need to take care about license notice because it is
already put in the source file.