add support for choosing between array/string/raw for the stderr or stdout data (#47)
One of the big mistakes I did in designing async.vim was to default to neovim output. Most of the time you need to get string and his can be costly for vim, since vim by default returns a string, which means we call split in async.vim then the consumer calls join which should had ideally been a noop. This allocates unnecessary memory due to string and array creation. Luckily we can have 100% backwards compatibility as I still default to 'array' if not passed and one can easily override it to 'string'. For callbag.vim's spawn function I had defaulted to 'raw' and make the user decide if they want compatibility.